+++ admin/install.php
@@ -62,6 +62,7 @@ $step = intval($_GET['step'] ?? 1);
  * @param array $data
  */
 function runInstall($data) {
+	// Extend the Query class
 	global $rs_query;
 	
 	// Include the database schema
@@ -83,8 +84,11 @@ function runInstall($data) {
 	// Populate the posts table
 	$home_page = populatePosts($author);
 	
+	// Get the site's url
+	$site_url = (!empty($_SERVER['HTTPS']) ? 'https://' : 'http://').$_SERVER['HTTP_HOST'];
+	
 	// Get the settings data
-	$settings = array('site_title'=>$data['site_title'], 'admin_email'=>$data['admin_email'], 'home_page'=>$home_page, 'do_robots'=>$data['do_robots']);
+	$settings = array('site_title'=>$data['site_title'], 'site_url'=>$site_url, 'admin_email'=>$data['admin_email'], 'home_page'=>$home_page, 'do_robots'=>$data['do_robots']);
 	
 	// Populate the settings table
 	populateSettings($settings);